Add files via upload
There has been a revenue decline in the Portuguese Bank, and they would like to know what actions to take. After investigation, they found that the root cause was that their customers were not investing enough for long-term deposits. So, the bank would like to identify existing customers with a higher chance of subscribing to a long-term deposit and focus marketing efforts on such customers.